Split different things up into separate files Tested it this time, it does work
